graphql: Apply hideViewerCursor and hideViewersAnnotation
