+++ bbb-graphql-middleware/internal/websrv/rediscli.go
@@ -1,11 +1,65 @@
-package rediscli
+package websrv
 
 import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/redis/go-redis/v9"
+	log "github.com/sirupsen/logrus"
+	"strings"
 	"time"
 )
+
+var redisClient = redis.NewClient(&redis.Options{
+	Addr:     "127.0.0.1:6379",
+	Password: "",
+	DB:       0,
+})
+
+func GetRedisConn() *redis.Client {
+	return redisClient
+}
+
+func StartRedisListener() {
+	log := log.WithField("_routine", "StartRedisListener")
+
+	var ctx = context.Background()
+
+	subscriber := GetRedisConn().Subscribe(ctx, "from-akka-apps-redis-channel")
+
+	for {
+		msg, err := subscriber.ReceiveMessage(ctx)
+		if err != nil {
+			log.Errorf("error: ", err)
+		}
+
+		// Skip parsing unnecessary messages
+		if !strings.Contains(msg.Payload, "InvalidateUserGraphqlConnectionSysMsg") {
+			continue
+		}
+
+		var message interface{}
+		if err := json.Unmarshal([]byte(msg.Payload), &message); err != nil {
+			panic(err)
+		}
+
+		messageAsMap := message.(map[string]interface{})
+
+		messageEnvelopeAsMap := messageAsMap["envelope"].(map[string]interface{})
+
+		messageType := messageEnvelopeAsMap["name"]
+
+		if messageType == "InvalidateUserGraphqlConnectionSysMsg" {
+			messageCoreAsMap := messageAsMap["core"].(map[string]interface{})
+			messageBodyAsMap := messageCoreAsMap["body"].(map[string]interface{})
+			sessionTokenToInvalidate := messageBodyAsMap["sessionToken"]
+			log.Debugf("Received invalidate request for sessionToken %v", sessionTokenToInvalidate)
+
+			//Not being used yet
+			InvalidateSessionTokenConnections(sessionTokenToInvalidate.(string))
+		}
+	}
+}
 
 func getCurrTimeInMs() int64 {
 	currentTime := time.Now()
