+++ bbb-graphql-server/metadata/databases/BigBlueButton/tables/public_v_meeting_lockSettings.yaml
@@ -18,6 +18,7 @@ select_permissions:
         - hasActiveLockSetting
         - hideUserList
         - hideViewersCursor
+        - hideViewersAnnotation
         - webcamsOnlyForModerator
       filter:
         meetingId:
