+++ bbb-graphql-server/metadata/databases/BigBlueButton/tables/public_v_pres_annotation_history_curr.yaml
@@ -6,6 +6,16 @@ configuration:
   custom_column_names: {}
   custom_name: pres_annotation_history_curr
   custom_root_fields: {}
+object_relationships:
+  - name: user
+    using:
+      manual_configuration:
+        column_mapping:
+          userId: userId
+        insertion_order: null
+        remote_table:
+          name: v_user_ref
+          schema: public
 select_permissions:
   - role: bbb_client
     permission:
@@ -17,5 +27,26 @@ select_permissions:
         - sequence
         - annotationInfo
       filter:
-        meetingId:
-          _eq: X-Hasura-MeetingId
+        _and:
+          - meetingId:
+              _eq: X-Hasura-MeetingId
+          - _or:
+              - meetingId:
+                  _eq: X-Hasura-ModeratorInMeeting
+              - userId:
+                  _eq: X-Hasura-LockedUserId
+              - meetingId:
+                  _neq: X-Hasura-LockedInMeeting
+              - _exists:
+                  _table:
+                    name: v_meeting_lockSettings
+                    schema: public
+                  _where:
+                    _and:
+                      - meetingId:
+                          _eq: X-Hasura-MeetingId
+                      - hideViewersAnnotation:
+                          _eq: false
+              - user:
+                  isModerator:
+                    _eq: true
