+++ bbb-graphql-server/metadata/databases/BigBlueButton/tables/public_v_pres_page_cursor.yaml
@@ -28,5 +28,26 @@ select_permissions:
         - xPercent
         - yPercent
       filter:
-        meetingId:
-          _eq: X-Hasura-MeetingId
+        _and:
+          - meetingId:
+              _eq: X-Hasura-MeetingId
+          - _or:
+              - user:
+                  isModerator:
+                    _eq: true
+              - meetingId:
+                  _eq: X-Hasura-ModeratorInMeeting
+              - meetingId:
+                  _neq: X-Hasura-LockedInMeeting
+              - userId:
+                  _eq: X-Hasura-LockedUserId
+              - _exists:
+                  _table:
+                    name: v_meeting_lockSettings
+                    schema: public
+                  _where:
+                    _and:
+                      - meetingId:
+                          _eq: X-Hasura-MeetingId
+                      - hideViewersCursor:
+                          _eq: false
