+++ bigbluebutton-web/grails-app/controllers/org/bigbluebutton/web/controllers/ConnectionController.groovy
@@ -84,6 +84,7 @@ class ConnectionController {
               "X-Hasura-Role" "bbb_client"
               "X-Hasura-Locked" u.locked ? "true" : "false"
               "X-Hasura-LockedInMeeting" u.locked ? userSession.meetingID : ""
+              "X-Hasura-LockedUserId" u.locked ? userSession.internalUserId : ""
               "X-Hasura-ModeratorInMeeting" u.isModerator() ? userSession.meetingID : ""
               "X-Hasura-PresenterInMeeting" u.isPresenter() ? userSession.meetingID : ""
               "X-Hasura-UserId" userSession.internalUserId
